fixed the issue of the SSE and fixed the issue of incomplete responses to client and added better UI
