--- PRODUCTION/src/recreate.html
@@ -1,145 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>ElixpoSearch SSE Demo</title>
-    <style>
-        body { font-family: Arial, sans-serif; margin: 2em; }
-        #output { white-space: pre-wrap; border: 1px solid #ccc; padding: 1em; min-height: 200px; }
-        #final { color: #1a8917; font-weight: bold; }
-        #step { color: #0057b8; }
-        #info { color: #888; }
-        #error { color: #b80000; }
-        .btn-group button { margin-right: 8px; }
-    </style>
-</head>
-<body>
-    <h2>ElixpoSearch API Test</h2>
-    <form id="queryForm">
-        <input type="text" id="queryInput" placeholder="Enter your query..." size="50" value="what's 1+1?" required>
-        <div class="btn-group">
-            <button type="button" id="btnGET">GET /search</button>
-            <button type="button" id="btnPOST">POST /search</button>
-            <button type="button" id="btnOpenAI">POST /search/v1/chat/completions</button>
-            <button type="button" id="btnSSE">POST /search/sse (SSE)</button>
-        </div>
-    </form>
-    <div id="output"></div>
-    <script>
-        const form = document.getElementById('queryForm');
-        const input = document.getElementById('queryInput');
-        const output = document.getElementById('output');
-        const btnGET = document.getElementById('btnGET');
-        const btnPOST = document.getElementById('btnPOST');
-        const btnOpenAI = document.getElementById('btnOpenAI');
-        const btnSSE = document.getElementById('btnSSE');
-
-        function showResult(label, data) {
-            output.innerHTML = '';
-            let div = document.createElement('div');
-            // Show error if present, else show result
-            if (typeof data === 'object' && data !== null && data.error) {
-                div.innerText = `[${label}] ERROR: ${data.error}`;
-                div.id = 'error';
-            } else if (typeof data === 'object' && data !== null && data.result !== undefined) {
-                div.innerText = `[${label}] ${data.result}`;
-            } else {
-                div.innerText = `[${label}] ${typeof data === 'string' ? data : JSON.stringify(data, null, 2)}`;
-            }
-            output.appendChild(div);
-        }
-
-        function testGET() {
-            output.innerHTML = 'Loading...';
-            fetch(`http://127.0.0.1:5000/search?query=${encodeURIComponent(input.value)}`)
-                .then(r => r.json())
-                .then(data => console.log(data))
-                .catch(e => showResult('GET', {error: e.toString()}));
-        }
-
-        function testPOST() {
-            output.innerHTML = 'Loading...';
-            fetch('http://127.0.0.1:5000/search', {
-                method: 'POST',
-                headers: { 'Content-Type': 'application/json' },
-                body: JSON.stringify({ query: input.value })
-            })
-            .then(r => r.json())
-            .then(data => showResult('POST', data))
-            .catch(e => showResult('POST', {error: e.toString()}));
-        }
-
-        function testOpenAI() {
-            output.innerHTML = 'Loading...';
-            fetch('http://127.0.0.1:5000/search/v1/chat/completions', {
-                method: 'POST',
-                headers: { 'Content-Type': 'application/json' },
-                body: JSON.stringify({
-                    messages: [
-                        { role: "user", content: input.value }
-                    ]
-                })
-            })
-            .then(r => r.json())
-            .then(data => {
-                let content = data?.choices?.[0]?.message?.content;
-                if (content !== undefined) {
-                    showResult('OpenAI', {result: content});
-                } else {
-                    showResult('OpenAI', data);
-                }
-            })
-            .catch(e => showResult('OpenAI', {error: e.toString()}));
-        }
-
-        function testSSE() {
-            output.innerHTML = '';
-            fetch('http://127.0.0.1:5000/search/sse', {
-                method: 'POST',
-                headers: { 'Content-Type': 'application/json' },
-                body: JSON.stringify({ query: input.value })
-            }).then(response => {
-                const reader = response.body.getReader();
-                let decoder = new TextDecoder();
-                let buffer = '';
-                function read() {
-                    reader.read().then(({ done, value }) => {
-                        if (done) return;
-                        buffer += decoder.decode(value, { stream: true });
-                        let parts = buffer.split('\n\n');
-                        buffer = parts.pop();
-                        for (let part of parts) {
-                            let lines = part.trim().split('\n');
-                            let event = '', data = '';
-                            for (let line of lines) {
-                                if (line.startsWith('event:')) event = line.replace('event:', '').trim();
-                                if (line.startsWith('data:')) data += line.replace('data:', '').trim() + '\n';
-                            }
-                            data = data.trim();
-                            let div = document.createElement('div');
-                            div.id = event;
-                            div.innerText = `[${event}] ${data}`;
-                            output.appendChild(div);
-                            if (event === 'final' || event === 'error') {
-                                output.appendChild(document.createElement('hr'));
-                            }
-                        }
-                        read();
-                    });
-                }
-                read();
-            });
-        }
-
-        btnGET.onclick = (e) => { e.preventDefault(); testGET(); };
-        btnPOST.onclick = (e) => { e.preventDefault(); testPOST(); };
-        btnOpenAI.onclick = (e) => { e.preventDefault(); testOpenAI(); };
-        btnSSE.onclick = (e) => { e.preventDefault(); testSSE(); };
-
-        // Default: GET on Enter
-        form.addEventListener('submit', function(e) {
-            e.preventDefault();
-            testGET();
-        });
-    </script>
-</body>
-</html>
